refactor: 상품 등록 훅 분리
